[core.room, generator] 367447: Transition from Choice point to transition point not possible https://bugs.eclipse.org/bugs/show_bug.cgi?id=367447
Now it is possible to connect a transition point via a chain of choice points with itself.
Still not allowed is to connect another state graph node with a transition point because this corresponds to a transition to history which can me modeled more clearly as a self transition in this case.